ml-service/requirements.txt
# FastAPI and server
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.6.1

# Machine Learning
scikit-learn==1.6.1
lightgbm==4.5.0
xgboost==2.1.3
numpy==2.2.1
pandas==2.2.3

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
python-multipart==0.0.20

# Monitoring and logging
prometheus-client==0.21.0

# Model persistence
joblib==1.4.2
pickle5==0.0.12

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.8.4
